+++ src/main/java/com/bluespurs/starterkit/api/OnlineStore.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.bluespurs.starterkit.data.ProductResult;
 
 public interface OnlineStore {
-	public List<ProductResult> searchProducts(String keyword, float min_price);
+	public List<ProductResult> searchProducts(String keyword, int min_price);
 
-	public ProductResult getCheapestMatch(String keyword, float min_price);
+	public ProductResult getCheapestMatch(String keyword, int min_price);
 }
